Созданы файлы data.js и util.js
